requirements.txt
python-telegram-bot
python-dotenv
faster-whisper
torch
mysql-connector-python
